requirements.txt
asgiref==3.7.2
Django==4.1.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.4.2
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
pymongo==3.12.3
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.2.4
typing_extensions==4.9.0
uritemplate==4.1.1
